Move addAgent() and waitForAgent() to agents.go
Also, make waitForAgent() a tiny bit more efficient by using
`getCurrentAgentsForNamespace()` in the first check.
