+++ pkg/client/connector/userd_trafficmgr/intercept.go
@@ -361,64 +361,6 @@ func (tm *trafficManager) AddIntercept(c context.Context, ir *rpc.CreateIntercep
 	}
 }
 
-func (tm *trafficManager) addAgent(c context.Context, namespace, agentName, svcName, svcPortIdentifier, agentImageName string) *rpc.InterceptResult {
-	svcUID, kind, err := tm.ensureAgent(c, namespace, agentName, svcName, svcPortIdentifier, agentImageName)
-	if err != nil {
-		if err == agentNotFound {
-			return &rpc.InterceptResult{
-				Error:     rpc.InterceptError_NOT_FOUND,
-				ErrorText: agentName,
-			}
-		}
-		dlog.Error(c, err)
-		return &rpc.InterceptResult{
-			Error:     rpc.InterceptError_FAILED_TO_ESTABLISH,
-			ErrorText: err.Error(),
-		}
-	}
-
-	dlog.Infof(c, "Waiting for agent for %s %s.%s", kind, agentName, namespace)
-	agent, err := tm.waitForAgent(c, agentName, namespace)
-	if err != nil {
-		dlog.Error(c, err)
-		return &rpc.InterceptResult{
-			Error:     rpc.InterceptError_FAILED_TO_ESTABLISH,
-			ErrorText: err.Error(),
-		}
-	}
-	dlog.Infof(c, "Agent found or created for %s %s.%s", kind, agentName, namespace)
-	return &rpc.InterceptResult{
-		Error:        rpc.InterceptError_UNSPECIFIED,
-		Environment:  agent.Environment,
-		ServiceUid:   svcUID,
-		WorkloadKind: kind,
-	}
-}
-
-func (tm *trafficManager) waitForAgent(ctx context.Context, name, namespace string) (*manager.AgentInfo, error) {
-	fullName := name + "." + namespace
-	waitCh := make(chan *manager.AgentInfo)
-	tm.agentWaiters.Store(fullName, waitCh)
-	defer tm.agentWaiters.Delete(fullName)
-
-	// Agent may already exist.
-	for _, agent := range tm.getCurrentAgents() {
-		if agent.Name == name && agent.Namespace == namespace {
-			return agent, nil
-		}
-	}
-
-	ctx, cancel := client.GetConfig(ctx).Timeouts.TimeoutContext(ctx, client.TimeoutAgentInstall) // installing a new agent can take some time
-	defer cancel()
-
-	select {
-	case <-ctx.Done():
-		return nil, client.CheckTimeout(ctx, fmt.Errorf("waiting for agent %q to be present", fullName))
-	case agent := <-waitCh:
-		return agent, nil
-	}
-}
-
 // shouldForward returns true if the intercept info given should result in mounts or ports being forwarded
 func (tm *trafficManager) shouldForward(ii *manager.InterceptInfo) bool {
 	return ii.SftpPort > 0 || len(ii.Spec.ExtraPorts) > 0
